This project focuses upon prediction of the release year of a particular song by using a song’s inherent timbre features. It will serve as a useful topic because people tend to have some affinity to different genres of music throughout their life.
Timbre is the quality of a musical note, sound, or tone that distinguishes different types of sound production, such as voices and musical
instruments This project focuses upon prediction of the release year of a
particular song by using a song’s inherent timbre features. It will serve as
a useful topic because people tend to have some affinity to different genres
of music throughout their life. This research will also help in determining
the long term evolution of music over the years and could be a useful basis
for recommending songs. The study could further be extended to generate new songs by learning the audio features of songs over the years. In
this research, we have applied four different predictive models on Million
Song Dataset and these are built on top of Apache Spark. The models
have been compared based on their accuracy to predict the release year as
well as their performance on a single node and multi-node Apache Spark
cluster environment.